requirements.txt
faiss-cpu==1.7.2
pandas==1.3.3
spacy==3.1.2
en-core-web-sm==3.1.0
google-generativeai

